fix the second caseo

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,23 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
+        User user= (User) session.getAttribute("loggeduser");
+        String error = "Only the owner of the image can edit the image";
 
         String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if(image.getUser().equals(user)) {
+            model.addAttribute("editError", error);
+            return "redirect:/images";
+        }
+        else {
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -114,7 +124,6 @@ public class ImageController {
     //The method converts the string to a list of all the tags using findOrCreateTags() method and sets the tags attribute of an image as a list of all the tags
     @RequestMapping(value = "/editImage", method = RequestMethod.PUT)
     public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
-        String error = "Only the owner of the image can edit the image";
         Image image = imageService.getImage(imageId);
 
         String updatedImageData = convertUploadedFileToBase64(file);
@@ -143,9 +152,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+        String error = "Only the owner of the image can edit the image";
+        Image image = imageService.getImage(imageId);
+        User user= (User) session.getAttribute("loggeduser");
+        if(image.getUser().equals(user)) {
+            model.addAttribute("deleteError", error);
+            return "/images/image";
+        }
+        else
+        {
         imageService.deleteImage(imageId);
-        return "redirect:/images";
+            return "redirect:/images";
+        }
+
     }
